DIO reads and writes working, not joystick yet.
I have a Teensy++ 2.0 now.  This code is tested working to read the Arcade buttons (ground other side to pull low when depressed), and write the LEDs in kind.   Not tested is the Joystick writes which don't seem to be detected on the Windows game controller properties so far.   Testing was frustrated by the fact that this Teensy seems to have a stuck DIO6 low.  I had to add extra look up tables to deal with skipping that DIO.